refactors - closes #8
