+++ components/Navbar/Navbar.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 
 import Login from './Login'
 import Logout from './Logout'
-import { Role, Status } from '../../server/lookups'
+import { Role, Status } from '../../shared/lookups'
 
 export default class Navbar extends Component {
