+++ shared/lookups.js
